Bug 1417 (JogAmp): New CCTask parameter 'writesoname': Adds GnuLinker and clang linker '-h name' options, writing SONAME for shares libraries
Setting the SONAME via general options simplifies the adoption of SONAME,
as the individual linker configurations must not change in the user projects.
This is new option for conveninience and scaling.

Feature is currently enabled for the GnuLinker (GccLinker 'gcc' + GppLinker 'gpp') as well as for clang (clang).

The SONAME must be set on Android API level >= 23,
see gluegen commit 51ef5eadd9db020412d3a3716b4ab5a25b0522fb.